Move spectrumfusion creation to end of fusion deployment
