+++ ocs_ci/ocs/constants.py
@@ -3201,7 +3201,9 @@ FDF_IMAGE_DIGEST_MIRROR_SET = os.path.join(
     FDF_TEMPLATE_DIR, "image-digest-mirror-set.yaml"
 )
 FDF_SERVICE_CR = os.path.join(FDF_TEMPLATE_DIR, "data-foundation-instance.yaml")
-FDF_SPECTRUM_FUSION_CR = os.path.join(FDF_TEMPLATE_DIR, "spectrum-fusion.yaml")
+SPECTRUM_FUSION_CR = os.path.join(
+    TEMPLATE_DEPLOYMENT_DIR_FUSION, "spectrum-fusion.yaml"
+)
 FDF_NAMESPACE = "ibm-spectrum-fusion-ns"
 ISF_CATALOG_SOURCE_NAME = "isf-catalog"
 ISF_OPERATOR_SOFTWARE_CATALOG_SOURCE_YAML = "catalog-source.yaml.j2"
