+++ ocs_ci/templates/fusion/spectrum-fusion.yaml
